+++ src/core/mouse/mouse.model.d.ts
@@ -1,0 +1,7 @@
+import { Td } from '../dom/td';
+
+export declare interface MouseModel {
+	code: 'left' | 'right' | 'middle';
+	status: 'release' | 'down' | 'up';
+	target: Td;
+}