+++ src/examples/column-reference-basic/example-column-reference-basic.component.ts
@@ -45,7 +45,7 @@ export class ExampleColumnReferenceBasicComponent {
 				})
 				.visibility({
 					toolbar: {
-						top: false,
+						top: true,
 						bottom: false,
 						right: false,
 						left: false