+++ src/examples/example.module.ts
@@ -46,6 +46,7 @@ import { ExampleDestroyGridModelComponent } from './destroy-grid-model/example-d
 import { ExampleDetailsRowAllComponent } from './details-row-all/example-details-row-all.component';
 import { ExampleDetailsRowApiComponent } from './details-row-api/example-details-row-api.component';
 import { ExampleDetailsRowBasicComponent } from './details-row-basic/example-details-row-basic.component';
+import { ExampleDetailsRowCustomComponent } from './details-row-custom/example-details-row-custom.component';
 import { ExampleDetailsRowGridComponent } from './details-row-grid/example-details-row-grid.component';
 import { ExampleDetailsRowPinComponent } from './details-row-pin/example-details-row-pin.component';
 import { ExampleDetailsRowStartComponent } from './details-row-start/example-details-row-start.component';
@@ -61,8 +62,9 @@ import { ExampleExportBasicComponent } from './export-basic/example-export-basic
 import { ExampleExportCsvComponent } from './export-csv/example-export-csv.component';
 import { ExampleFilterColumnFetchComponent } from './filter-column-fetch/example-filter-column-fetch.component';
 import { ExampleFilterConditionBasicComponent } from './filter-condition-basic/example-filter-condition-basic.component';
-import { ExampleFilterRowBasicComponent } from './filter-row-basic/example-filter-row-basic.component';
+import { ExampleFilterRowAtomBasicComponent } from './filter-row-atom-basic/example-filter-row-atom-basic.component';
 import { ExampleFilterRowCustomComponent } from './filter-row-custom/example-filter-row-custom.component';
+import { ExampleFilterRowPeopleBasicComponent } from './filter-row-people-basic/example-filter-row-people-basic.component';
 import { ExampleFloatingRowsBasicComponent } from './floating-rows-basic/example-floating-rows-basic.component';
 import { ExampleFocusCellAutoComponent } from './focus-cell-auto/example-focus-cell-auto.component';
 import { ExampleFocusCellBasicComponent } from './focus-cell-basic/example-focus-cell-basic.component';
@@ -184,7 +186,7 @@ export const EXAMPLES: any[] = [
 	ExampleColumnUrlBasicComponent,
 	ExampleDataRowAddComponent,
 	ExampleDataRowDeleteComponent,
-	ExampleDefineColumnAsyncComponent, 
+	ExampleDefineColumnAsyncComponent,
 	ExampleDefineColumnBasicComponent,
 	ExampleDefineColumnHybridComponent,
 	ExampleDefineColumnTypeComponent,
@@ -193,6 +195,7 @@ export const EXAMPLES: any[] = [
 	ExampleDetailsRowAllComponent,
 	ExampleDetailsRowApiComponent,
 	ExampleDetailsRowBasicComponent,
+	ExampleDetailsRowCustomComponent,
 	ExampleDetailsRowGridComponent,
 	ExampleDetailsRowPinComponent,
 	ExampleDetailsRowStartComponent,
@@ -208,8 +211,9 @@ export const EXAMPLES: any[] = [
 	ExampleExportCsvComponent,
 	ExampleFilterColumnFetchComponent,
 	ExampleFilterConditionBasicComponent,
-	ExampleFilterRowBasicComponent,
+	ExampleFilterRowAtomBasicComponent,
 	ExampleFilterRowCustomComponent,
+	ExampleFilterRowPeopleBasicComponent,
 	ExampleFloatingRowsBasicComponent,
 	ExampleFocusCellAutoComponent,
 	ExampleFocusCellBasicComponent,