+++ src/lib/plugins/pane/pane.component.ts
@@ -16,12 +16,7 @@ export class PaneComponent implements OnInit {
 
 	context: {
 		$implicit: PaneComponent,
-		cell: {
-			rowIndex: number,
-			columnIndex: number,
-			row: any,
-			column: any
-		};
+		value: any;
 	};
 
 	constructor(
@@ -34,13 +29,9 @@ export class PaneComponent implements OnInit {
 
 	ngOnInit() {
 		const { model } = this.plugin;
-		const parts = this.trigger ? this.trigger.split('.') : [];
-		if (parts.length > 0) {
-			const [state, prop] = parts;
-			if (!model[state]) {
-				throw new AppError('pane.component', `Trigger ${state} not found`);
-			}
-
+		const scope = this.parse();
+		if (scope) {
+			const [state, prop] = scope;
 			model[`${state}Changed`].watch(e => {
 				if (!prop || e.hasChanges(prop)) {
 					this.close('right');
@@ -53,12 +44,16 @@ export class PaneComponent implements OnInit {
 	open(side: 'right') {
 		const { table, model } = this.plugin;
 
-		this.context = {
-			$implicit: this,
-			cell: model.navigation().cell
-		};
+		let value = null;
+		const scope = this.parse();
+		if (scope) {
+			const [state, prop] = scope;
+			value = model[state]()[prop];
+		}
 
+		this.context = { $implicit: this, value };
 		this.cd.markForCheck();
+
 		table.view.addLayer(`pane-${side}`);
 	}
 
@@ -69,5 +64,20 @@ export class PaneComponent implements OnInit {
 
 		this.context = null;
 		this.cd.markForCheck();
+	}
+
+	private parse() {
+		const { model } = this.plugin;
+		const parts = this.trigger ? this.trigger.split('.') : [];
+		if (parts.length > 0) {
+			const [state, prop] = parts;
+			if (!model[state]) {
+				throw new AppError('pane.component', `Trigger ${state} not found`);
+			}
+
+			return [state, prop];
+		}
+
+		return null;
 	}
 }